Handle DTB-Id with a subdirectory
In v3.19-rc1 Linux arm64 moved all DTB files into per-vendor subdirectories and
the Debian kernel package has followed suite.

Update the entry for Mustang (the only arm64 platform in the current
flash-kernel database) to add the apm/ prefix and adjust the code which
installs and removes the dtb to cope with subdirectories in this context.